+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version latest (0.5.0?, 2019-09-24?)
 - bugfix: matomo ping was being called from test, thus cluttering the stats with noise. Mocked out
 - enh: `isitfit version` and `isitfit --version` both show the version (in case a user uses any of the 2)
 - enh: move all cost-related code into `isitfit/cost`
+- enh: move all tags-related code into `isitfit/tags`
 
 
 Version 0.10.2 (2019-10-11)